requirements.txt
asgiref==3.4.1
Django==3.2.4
django-cors-headers==3.7.0
django-mysql==3.11.1
djangorestframework==3.12.2
Faker==8.9.1
mysqlclient==2.0.3
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
